Find project namespaces before listing and filter
Now, identifies and supplies namespaces to namespace filter for requests
associated with given project. Prior, the backend listed items belonging
to all namespaces in all projects no matter what, and then filtered. This
led to longer request times for projects that contained a fraction of
total resources. Now, request times are improved and largely dependent on
resource amount for the given project.